chore(core): backend restructuration (9)
