+++ src/bp/core/app/botpress.ts
@@ -1,25 +1,25 @@
 import * as sdk from 'botpress/sdk'
 import lang from 'common/lang'
-import { createForGlobalHooks } from 'core/api'
+import { createForGlobalHooks } from 'core/app/api'
 import { BotService, BotMonitoringService } from 'core/bots'
 import { GhostService } from 'core/bpfs'
 import { CMSService } from 'core/cms'
 import { BotpressConfig, ConfigProvider } from 'core/config'
 import { buildUserKey, converseApiEvents } from 'core/converse'
 import Database from 'core/database'
-import { StateManager, DecisionEngine, DialogEngine, DialogJanitor } from 'core/dialog'
+import { StateManager, DecisionEngine, DialogEngine, DialogJanitor, WellKnownFlags } from 'core/dialog'
 import { SessionIdFactory } from 'core/dialog/sessions'
-import { addStepToEvent, EventCollector, StepScopes, StepStatus, EventEngine } from 'core/events'
+import { addStepToEvent, EventCollector, StepScopes, StepStatus, EventEngine, Event } from 'core/events'
 import { AlertingService, MonitoringService } from 'core/health'
 import { LoggerDbPersister, LoggerFilePersister, LoggerProvider, LogsJanitor } from 'core/logger'
 import { MigrationService } from 'core/migration'
 import { copyDir } from 'core/misc/pkg-fs'
 import { ModuleLoader } from 'core/modules'
 import { NotificationsService } from 'core/notifications'
+import { RealtimeService } from 'core/realtime'
 import { AuthService } from 'core/security'
 import { StatsService } from 'core/telemetry'
-import { TYPES } from 'core/types'
-import { ActionServersConfigSchema, Hooks, HookService } from 'core/user-code'
+import { ActionServersConfigSchema, Hooks, HookService, HintsService } from 'core/user-code'
 import { DataRetentionJanitor, DataRetentionService, WorkspaceService } from 'core/users'
 import { WrapErrorsWith } from 'errors'
 import fse from 'fs-extra'
@@ -33,14 +33,12 @@ import nanoid from 'nanoid'
 import path from 'path'
 import plur from 'plur'
 
-import { startLocalActionServer } from '../cluster'
-import { setDebugScopes } from '../debug'
-import { WellKnownFlags } from './sdk/enums'
-import { Event } from './sdk/impl'
-import HTTPServer from './server'
-import { HintsService } from './services/hints'
-import RealtimeService from './services/realtime'
-import { Statistics } from './stats'
+import { startLocalActionServer } from '../../cluster'
+import { setDebugScopes } from '../../debug'
+
+import { Statistics } from '../stats'
+import { HTTPServer } from './server'
+import { TYPES } from './types'
 
 export interface StartOptions {
   modules: sdk.ModuleEntryPoint[]
