+++ src/bp/core/converse/converse-router.ts
@@ -1,8 +1,8 @@
 import { Logger } from 'botpress/sdk'
 import { StandardError } from 'common/http'
+import { HTTPServer } from 'core/app/server'
 import { ConverseService } from 'core/converse'
 import { AuthService, TOKEN_AUDIENCE, checkTokenHeader } from 'core/security'
-import HTTPServer from 'core/server'
 import { RequestHandler, Router } from 'express'
 import joi from 'joi'
 import _ from 'lodash'
