+++ src/bp/core/dialog/enums.ts
@@ -1,19 +1,5 @@
 import { IO } from 'botpress/sdk'
 
-export enum LoggerLevel {
-  Info = 'info',
-  Warn = 'warn',
-  Error = 'error',
-  Critical = 'critical',
-  Debug = 'debug'
-}
-
-export enum LogLevel {
-  PRODUCTION = 0,
-  DEV = 1,
-  DEBUG = 2
-}
-
 export enum NodeActionType {
   RenderElement = 'render',
   RunAction = 'run',
