+++ src/bp/core/events/event-engine.ts
@@ -1,16 +1,16 @@
 import * as sdk from 'botpress/sdk'
+import { TYPES } from 'core/app/types'
+import { WellKnownFlags } from 'core/dialog'
 import { incrementMetric } from 'core/health'
 import { TimedPerfCounter } from 'core/misc/timed-perf'
-import { WellKnownFlags } from 'core/sdk/enums'
-import { TYPES } from 'core/types'
 import { inject, injectable, tagged } from 'inversify'
 import joi from 'joi'
 import _ from 'lodash'
 import { VError } from 'verror'
 import yn from 'yn'
 
-import { Event } from '../sdk/impl'
 import { EventCollector } from './event-collector'
+import { Event } from './event-sdk-impl'
 import { MiddlewareChain } from './middleware-chain'
 import { Queue } from './queue'
 import { addStepToEvent, StepScopes } from './utils'
