+++ src/bp/core/events/index.ts
@@ -1,6 +1,7 @@
 export * from './event-collector'
 export * from './event-engine'
 export * from './event-repository'
+export * from './event-sdk-impl'
 export * from './queue'
 export * from './queue/memory-queue'
 export * from './utils'
