+++ src/bp/core/messaging/message-service.ts
@@ -1,10 +1,9 @@
 import * as sdk from 'botpress/sdk'
+import { TYPES } from 'core/app/types'
 import { JobService } from 'core/distributed'
-import { EventEngine } from 'core/events'
+import { EventEngine, IOEvent } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
 import { MessageRepository } from 'core/messaging'
-import { IOEvent } from 'core/sdk/impl'
-import { TYPES } from 'core/types'
 import { inject, injectable, postConstruct } from 'inversify'
 import { AppLifecycle, AppLifecycleEvents } from 'lifecycle'
 import LRU from 'lru-cache'
