+++ src/bp/core/realtime/index.ts
@@ -1,3 +1,6 @@
+export * from './payload-sdk-impl'
+
 // TODO: Move code here when references are updated
 
-export * from '../services/realtime'
+import RealtimeService from '../services/realtime'
+export { RealtimeService }
