+++ src/bp/core/services/auth/auth-service.ts
@@ -1,16 +1,15 @@
 import { Logger, StrategyUser } from 'botpress/sdk'
 import { JWT_COOKIE_NAME } from 'common/auth'
+import { TYPES } from 'core/app/types'
 import { AuthStrategy, ConfigProvider } from 'core/config'
 import Database from 'core/database'
 import { SessionIdFactory } from 'core/dialog/sessions'
 import { JobService } from 'core/distributed'
-import { EventEngine } from 'core/events'
+import { EventEngine, Event } from 'core/events'
 import { KeyValueStore } from 'core/kvs'
 import { ModuleLoader } from 'core/modules'
 import { BadRequestError } from 'core/routers/errors'
-import { Event } from 'core/sdk/impl'
 import { StrategyBasic, generateUserToken, getMessageSignature } from 'core/security'
-import { TYPES } from 'core/types'
 import { StrategyUsersRepository, WorkspaceService } from 'core/users'
 import { StrategyUserTable } from 'core/users/tables'
 import { Response } from 'express'
