+++ src/bp/core/services/bot-service.ts
@@ -3,6 +3,7 @@ import cluster from 'cluster'
 import { BotHealth, ServerHealth } from 'common/typings'
 import { BotCreationSchema, BotEditSchema, isValidBotId } from 'common/validation'
 import { createForGlobalHooks } from 'core/api'
+import { TYPES } from 'core/app/types'
 import { FileContent, GhostService, ReplaceContent } from 'core/bpfs'
 import { CMSService } from 'core/cms'
 import { ConfigProvider } from 'core/config'
@@ -13,9 +14,8 @@ import { IDisposable } from 'core/misc/disposable'
 import { listDir } from 'core/misc/list-dir'
 import { stringify } from 'core/misc/utils'
 import { ModuleResourceLoader, ModuleLoader } from 'core/modules'
-import { RealTimePayload } from 'core/sdk/impl'
+import { RealtimeService, RealTimePayload } from 'core/realtime'
 import { Statistics } from 'core/stats'
-import { TYPES } from 'core/types'
 import { Hooks, HookService } from 'core/user-code'
 import { WorkspaceService } from 'core/users'
 import { WrapErrorsWith } from 'errors'
@@ -33,10 +33,7 @@ import tmp from 'tmp'
 import { VError } from 'verror'
 
 import { extractArchive } from '../misc/archive'
-
 import { InvalidOperationError } from './auth/errors'
-
-import RealtimeService from './realtime'
 
 const BOT_DIRECTORIES = ['actions', 'flows', 'entities', 'content-elements', 'intents', 'qna']
 const BOT_CONFIG_FILENAME = 'bot.config.json'
