+++ src/bp/core/services/dialog/decision-engine.ts
@@ -1,9 +1,8 @@
 import { IO, Logger, NDU } from 'botpress/sdk'
+import { TYPES } from 'core/app/types'
 import { ConfigProvider } from 'core/config'
-import { StateManager } from 'core/dialog'
+import { StateManager, WellKnownFlags } from 'core/dialog'
 import { EventEngine } from 'core/events'
-import { WellKnownFlags } from 'core/sdk/enums'
-import { TYPES } from 'core/types'
 import { inject, injectable, postConstruct, tagged } from 'inversify'
 import { AppLifecycle, AppLifecycleEvents } from 'lifecycle'
 import _ from 'lodash'
