+++ src/bp/core/services/dialog/flow/service.ts
@@ -2,11 +2,12 @@ import { Flow, Logger } from 'botpress/sdk'
 import { ObjectCache } from 'common/object-cache'
 import { TreeSearch, PATH_SEPARATOR } from 'common/treeSearch'
 import { FlowMutex, FlowView, NodeView } from 'common/typings'
+import { TYPES } from 'core/app/types'
+import { BotService } from 'core/bots'
+import { GhostService } from 'core/bpfs'
 import { KeyValueStore } from 'core/kvs'
 import { ModuleLoader } from 'core/modules'
-import { RealTimePayload } from 'core/sdk/impl'
-import { BotService } from 'core/services/bot-service'
-import RealtimeService from 'core/services/realtime'
+import { RealtimeService, RealTimePayload } from 'core/realtime'
 import { inject, injectable, tagged } from 'inversify'
 import Joi from 'joi'
 import _ from 'lodash'
@@ -14,8 +15,6 @@ import { Memoize } from 'lodash-decorators'
 import moment from 'moment'
 import nanoid from 'nanoid/generate'
 
-import { GhostService } from '../..'
-import { TYPES } from '../../../types'
 import { validateFlowSchema } from '../validator'
 
 const PLACING_STEP = 250
