+++ src/bp/core/user-code/index.ts
@@ -4,3 +4,5 @@ export * from './action-server/tasks-repository'
 export * from './action-service'
 export * from './hook-service'
 export * from './vm'
+export * from './hints/hints-service'
+export * from './hints/hints-router'
