+++ src/bp/diag/index.ts
@@ -6,7 +6,7 @@ import '../sdk/rewire'
 import { BotConfig } from 'botpress/sdk'
 
 import { Workspace } from 'common/typings'
-import { Db, Ghost } from 'core/app'
+import { Db, Ghost } from 'core/app/core-loader'
 import { getOrCreate as redisFactory } from 'core/services/redis'
 import fse from 'fs-extra'
 import IORedis from 'ioredis'
