+++ src/bp/nlu/stan/index.ts
@@ -5,15 +5,13 @@ import rewire from '../../sdk/rewire'
 import bytes from 'bytes'
 import chalk from 'chalk'
 import cluster from 'cluster'
-
+import { LogLevel } from 'core/logger/enums'
 import { centerText } from 'core/logger/utils'
 import { copyDir } from 'core/misc/pkg-fs'
 import _ from 'lodash'
 import * as NLUEngine from 'nlu/engine'
 import path from 'path'
-
 import { setupMasterNode, WORKER_TYPES } from '../../cluster'
-import { LogLevel } from '../../core/sdk/enums'
 
 import Logger from '../../simple-logger'
 import API, { APIOptions } from './api'
